refactor: useResult 커스텀훅 분리
